+++ src/blockchain/leveldb/leveldb_organizer.hpp
@@ -37,8 +37,8 @@ public:
         leveldb_chain_keeper_ptr chain, reorganize_handler handler);
 
 protected:
-    std::error_code verify(int fork_index,
-        const block_detail_list& orphan_chain, int orphan_index);
+    std::error_code verify(size_t fork_index,
+        const block_detail_list& orphan_chain, size_t orphan_index);
     void reorganize_occured(
         size_t fork_point,
         const blockchain::block_list& arrivals,
